services/requirements.txt
langchain
requests
fitz
PyMuPDF
langchain-google-genai
pdf2image
pytesseract
python-dotenv
schedule
feedparser
